pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "inkscape-shortcuts"
version = "0.1.0"
dependencies = [
    "python-xlib", # We will need this for the X11 interception
]

# [project.scripts]
